# openapi/partner-scheduling.yaml
openapi: 3.1.0
info:
  title: Rula Partner API
  version: 0.23.2
  description: >
    Rula's API for partners. This API is currently a WIP and may change at any time.

    Updates:
    v0.23.2
      Changelog:
        + Clarified start_time_range description on GET /v1/providers/slots
        + Improved descriptions across multiple fields for clarity
        + Added minItems/maxItems: 2 constraint to start_time_range on GET /v1/providers/slots

    v0.23.1
      Changelog:
        + Removed request-start-time header parameter from spec (now handled at the gateway level)

    v0.23.0
      Changelog:
        + Added optional is_self_pay boolean field to PatientCreateRequest (default: false)
        + insurance_id is now only required when is_self_pay is false
        + subscriber_id is now only required when is_eap_referral is false and is_self_pay is false
        + is_eap_referral must be false when is_self_pay is true

    v0.22.0
      Changelog:
        + Added PUT /v1/appointments/{uuid}/cancel endpoint to cancel an appointment

    v0.21.1
      Changelog:
        + Added request-start-time as an accepted optional header on all endpoints

    v0.21.0
      Changelog:
        + Added GET /v1/appointments/{uuid} endpoint to retrieve a single appointment
        + Extracted shared Appointment schema used by PatientAppointmentsResponse and AppointmentShowResponse

    v0.20.1
      Changelog:
...

    v0.20.0
      Changelog:
...

    v0.19.0
      Changelog:
...

    v0.18.0
...

...

...

...

...

...

...

...

...

...
servers:
...
paths:
...
  /v1/providers/{uuid}:
    get:
      operationId: providersShow
      summary: Information for a specific provider.
      security:
        - auth0:
            - partner_api:providers:read
      parameters:
        - schema:
            $ref: "#/components/schemas/Uuid"
          example: 9c823227-dbe1-408c-9e71-0685817e3410
          in: path
          name: uuid
          required: true
          description: The uuid of the provider to retrieve
      responses:
        "200":
          $ref: "#/components/responses/ProvidersShowResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestValidationError"
        "401":
          description: Unauthorized, invalid or expired token
        "422":
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestValidationError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
  /v1/providers/search:
    post:
      operationId: providersSearch
      summary: Search for providers matching criteria, returns summary information.
      security:
        - auth0:
            - partner_api:providers:read
      requestBody:
        $ref: "#/components/requestBodies/ProvidersSearchRequest"
      responses:
        "200":
          $ref: "#/components/responses/ProvidersSearchResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestValidationError"
        "401":
          description: Unauthorized, invalid or expired token
        "422":
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestValidationError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
  /v1/providers/slots:
    get:
      operationId: providersSlots
      summary: Available slots for a specific provider.
      security:
        - auth0:
            - partner_api:providers:read
      parameters:
        - schema:
            $ref: "#/components/schemas/Uuid"
          in: query
          name: provider_uuid
          required: true
          description: The unique Rula-provided identifier for the provider.
        - schema:
            type: string
            pattern: ^[A-Z]{2}$
            maxLength: 2
          in: query
          name: two_letter_state
          required: true
          description: Two-letter state code (e.g., "CA" for California).
        - schema:
            type: array
            minItems: 2
            maxItems: 2
            items:
              type: string
              format: date-time
          in: query
          name: start_time_range
          required: false
          description: A [start, end] pair of ISO 8601 datetime strings defining the window to fetch slots within. The first element is the start (inclusive) and the second is the end (inclusive). Filters by slot start time. When omitted, defaults to 25 hours from now through 28 days out. No maximum window is enforced.
        - schema:
            type: array
            uniqueItems: true
            minItems: 1
            items:
              type: string
              enum:
                - in_person
                - telemedicine
          in: query
          name: location_type
          required: true
          style: form
          explode: true
          description: Array of location types to filter appointment slots by (in-person vs virtual).
      responses:
        "200":
          $ref: "#/components/responses/ProvidersSlotsResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestValidationError"
        "401":
          description: Unauthorized, invalid or expired token
        "422":
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestValidationError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerError"
...
components:
  requestBodies:
...
    ProvidersSearchRequest:
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              limit:
                description: The maximum number of providers to return, between 1 and 50
                type: integer
                minimum: 1
                maximum: 50
                default: 50
              two_letter_state:
                description: "Required. The two-letter state the provider is credentialed in (e.g., \"CA\", \"NY\")."
                type: string
                pattern: ^[A-Z]{2}$
              care_category:
                description: The category of care to filter providers by. "therapy" returns therapists; "psychiatry" returns prescribers. Defaults to "therapy" if omitted.
                type: string
                enum:
                  - therapy
                  - psychiatry
                default: therapy
              insurance:
                description: Insurance network accepted by the provider. Must match the
                  network_name value returned from the GET /insurances
                  endpoint. The value is case-insensitive.
                type: string
              coverage_type:
                description: The coverage type of the insurance accepted by the provider
                type: string
                enum:
                  - commercial
                  - medicare_advantage
                  - medicare_ffs
                  - medicaid
                  - HMO
                  - eap
              gender:
                description: The genders the provider identifies with
                type: string
                enum:
                  - Agender
                  - Bigender
                  - Cis
                  - Female
                  - Gender fluid
                  - Male
                  - My gender isn't listed
                  - Non-binary
                  - Prefer not to respond
                  - Trans
              race:
                description: The races or ethnicities the provider identifies with
                type: string
                enum:
                  - American Indian or Alaska Native
                  - Asian
                  - Biracial or Multiracial
                  - Black or African American
                  - Caucasian or White
                  - Hispanic or Latinx
                  - Middle Eastern
                  - Native Hawaiian or Other Pacific Islander
                  - Other
                  - Prefer Not to Respond
                  - South East Asian
              language:
                description: The languages spoken by the provider
                type: string
                enum:
                  - Afrikaans
                  - Arabic
                  - Armenian
                  - ASL
                  - Bengali
                  - Bosnian
                  - Cantonese
                  - Chinese
                  - Czech
                  - Dutch
                  - English
                  - Farsi
                  - Filipino
                  - French
                  - German
                  - Greek
                  - Haitian Creole
                  - Hebrew
                  - Hindi
                  - Hmong
                  - Hungarian
                  - Indonesian
                  - Iranian
                  - Italian
                  - Japanese
                  - Javanese
                  - Korean
                  - My selection isn't listed
                  - Norwegian
                  - Other
                  - Patois
                  - Polish
                  - Portuguese
                  - Punjabi
                  - Romanian
                  - Russian
                  - Spanish
                  - Swahili
                  - Swedish
                  - Thai
                  - Turkish
                  - Ukrainian
                  - Urdu
                  - Vietnamese
                  - Yoruba
              specialization:
                description: The specializations the provider has
                type: string
                enum:
                  - Addiction
                  - ADHD
                  - Adoption
                  - Aging
                  - Alcohol Use
                  - Alzheimer's
                  - Anger Management
                  - Antisocial Personality
                  - Anxiety
                  - Autism or Asperger's Syndrome
                  - Behavioral Issues
                  - Bipolar Disorder
                  - Borderline Personality
                  - Career Counseling
                  - Child Abuse or Neglect
                  - Chronic Illness
                  - Chronic Impulsivity
                  - Chronic Pain
                  - Chronic Relapse
                  - Codependency
                  - Coping Skills
                  - Cultural Stress
                  - Depersonalization/Derealization (DPDR)
                  - Depression
                  - Developmental Disorders
                  - Dissociative Disorders
                  - Divorce
                  - Domestic Violence
                  - Drug Abuse
                  - Eating Disorders
                  - Family Conflict
                  - Gambling
                  - Gender Identity
                  - Grief
                  - Hoarding
                  - Impulse Control Disorders
                  - Impulsivity
                  - Infertility
                  - Infidelity
                  - Intellectual Disability
                  - Internet Addiction
                  - Job Stress
                  - Learning Disabilities
                  - Legal Stress
                  - LGBTQIA+
                  - Life Transitions
                  - Men's Issues
                  - Military and Veterans
                  - Mood Disorders
                  - Obsessive-Compulsive (OCD)
                  - Older Adult
                  - Older Adults
                  - Oppositional Defiance
                  - Panic Attacks
                  - Parenting
                  - Peer Relationships
                  - Personality Disorders
                  - Pregnancy, Prenatal, Postpartum
                  - Psychosis
                  - Racial Identity
                  - Racial Trauma
                  - Relapse Prevention
                  - Relationship Issues
                  - Schizophrenia
                  - School Issues
                  - Self Esteem
                  - Self-Harming
                  - Sex Therapy
                  - Sexual Abuse
                  - Sexual Addiction
                  - Sleep or Insomnia
                  - Spirituality
                  - Sports Performance
                  - Stress
                  - Substance Use
                  - Suicidal Ideation
                  - Thought Disorders
                  - Transgender Issues
                  - Trauma and PTSD
                  - Traumatic Brain Injury
                  - Video Game Addiction
                  - Weight Loss
                  - Women's Issues
              modality:
                description: The therapeutic modalities the provider practices
                type: string
                enum:
                  - Accelerated Experiential Dynamic Psychotherapy
                  - Accelerated Experiential Dynamic Psychotherapy (AEDP)
                  - Acceptance and Commitment (ACT)
                  - Art Therapy
                  - Attachment-based
                  - Behavioral Activation
                  - Biofeedback
                  - Brainspotting
                  - Child/Parent Psychotherapy
                  - Christian Counseling
                  - Cognitive Behavioral (CBT)
                  - Cognitive Processing (CPT)
                  - Cognitive Processing Therapy (CPT)
                  - Collaborative Assessment and Management of Suicidality
                    (CAMS)
                  - Collaborative Assessment & Management of Suicidality (CAMS)
                  - Compassion Focused
                  - Compassion Focused Therapy
                  - Complicated Grief Therapy (CGT)
                  - Dance/Movement Therapy
                  - Dialectical (DBT)
                  - EMDR
                  - Emotionally Focused Therapy (EFT)
                  - Existential Therapy
                  - Experiential Therapy
                  - Exposure and Response Prevention (ERP)
                  - Exposure Response Prevention
                  - Expressive Arts
                  - Gestalt
                  - Gottman Method
                  - Humanistic
                  - Humanistic Therapy
                  - Imago Relationship Therapy
                  - Internal Family Systems (IFS)
                  - Interpersonal
                  - Interpersonal Therapy (IPT)
                  - Mindfulness-Based
                  - Moral Reconation Therapy (MRT)
                  - Motivational Interviewing
                  - Multidimensional Family Therapy
                  - Multidimensional Family Therapy (MDFT)
                  - Music Therapy
                  - Narrative
                  - Narrative Therapy
                  - Neurofeedback
                  - Object Relations
                  - Parent-Child Interaction (PCIT)
                  - Parent-Child Interaction Therapy (PCIT)
                  - Play Therapy
                  - Prolonged Exposure (PE)
                  - Prolonged Exposure Therapy
                  - Psychoanalytic
                  - Psychobiological Approach Couples Therapy (PACT)
                  - Psychobiological Approach to Couples Therapy (PACT)
                  - Psychodynamic
                  - Psychodynamic Therapy
                  - Rational Emotive Behavior (REBT)
                  - Rational Emotive Behavior Therapy (REBT)
                  - Reality Therapy
                  - Sandplay
                  - Schema Therapy
                  - Sex Therapy
                  - Solution Focused Brief (SFBT)
                  - Somatic
                  - Somatic Therapy
                  - Structural Family Therapy
                  - Trauma Focused
                  - Trauma Focused Dialectical (DBT)
                  - Trust-Based Relational Intervention
                  - Trust-Based Relational Intervention (TBRI)
              willing_to_see:
                description: Criteria of patients the provider is willing to see
                oneOf:
                  - const: adults
                    description: Adults (18+)
                  - const: children
                    description: Children (ages 5-7)
                  - const: minors
                    description: Minors (ages 13-17)
                  - const: preteen
                    description: Preteen (ages 8-12)
                  - const: individuals
                    description: Individual therapy
                  - const: couples
                    description: Couples therapy
                  - const: families
                    description: Family therapy
              location_type:
                description: The types of locations where the provider offers service
                type: string
                enum:
                  - in_person
                  - telemedicine
              proximity:
                description: Filter providers by distance from a geographic location. Useful for finding in-person providers near a patient's address.
                type: object
                properties:
                  lat:
                    description: Latitude of the search center
                    type: number
                  long:
                    description: Longitude of the search center
                    type: number
                  max_miles:
                    description: Maximum distance in miles from the search center
                    type: integer
                    exclusiveMinimum: 0
                    maximum: 9007199254740991
                required:
                  - lat
                  - long
                  - max_miles
              availability:
                description: Filter providers to those who have availability within the specified weekly schedule (e.g. weekday mornings). All days of the week and time_zone are required. Cannot be specified if absolute_availability is provided.
                type: object
                properties:
                  sun:
                    description: An array of hour spans (start hour 0-23, end hour 1-24)
                    maxItems: 4
                    type: array
                    items:
                      type: array
                      items:
                        type: number
                        minimum: 0
                        maximum: 24
                      minItems: 2
                      maxItems: 2
                  mon:
                    description: An array of hour spans (start hour 0-23, end hour 1-24)
                    maxItems: 4
                    type: array
                    items:
                      type: array
                      items:
                        type: number
                        minimum: 0
                        maximum: 24
                      minItems: 2
                      maxItems: 2
                  tue:
                    description: An array of hour spans (start hour 0-23, end hour 1-24)
                    maxItems: 4
                    type: array
                    items:
                      type: array
                      items:
                        type: number
                        minimum: 0
                        maximum: 24
                      minItems: 2
                      maxItems: 2
                  wed:
                    description: An array of hour spans (start hour 0-23, end hour 1-24)
                    maxItems: 4
                    type: array
                    items:
                      type: array
                      items:
                        type: number
                        minimum: 0
                        maximum: 24
                      minItems: 2
                      maxItems: 2
                  thu:
                    description: An array of hour spans (start hour 0-23, end hour 1-24)
                    maxItems: 4
                    type: array
                    items:
                      type: array
                      items:
                        type: number
                        minimum: 0
                        maximum: 24
                      minItems: 2
                      maxItems: 2
                  fri:
                    description: An array of hour spans (start hour 0-23, end hour 1-24)
                    maxItems: 4
                    type: array
                    items:
                      type: array
                      items:
                        type: number
                        minimum: 0
                        maximum: 24
                      minItems: 2
                      maxItems: 2
                  sat:
                    description: An array of hour spans (start hour 0-23, end hour 1-24)
                    maxItems: 4
                    type: array
                    items:
                      type: array
                      items:
                        type: number
                        minimum: 0
                        maximum: 24
                      minItems: 2
                      maxItems: 2
                  time_zone:
                    description: IANA timezone name (e.g., 'America/Los_Angeles')
                    type: string
                required:
                  - sun
                  - mon
                  - tue
                  - wed
                  - thu
                  - fri
                  - sat
                  - time_zone
              absolute_availability:
                description: An array of [start, end] ISO 8601 datetime pairs representing specific date-time windows the provider must be available within. Cannot be specified if availability is provided.
                type: array
                items:
                  minItems: 2
                  maxItems: 2
                  type: array
                  items:
                    type: string
            required:
              - two_letter_state
          example:
            two_letter_state: CA
            insurance: Aetna
            availability:
              sun:
                - - 0
                  - 8
                - - 16
                  - 20
              mon:
                - - 9
                  - 17
              tue:
                - - 9
                  - 17
              wed:
                - - 9
                  - 17
              thu:
                - - 9
                  - 17
              fri:
                - - 9
                  - 17
              sat: []
              time_zone: America/Los_Angeles
            absolute_availability:
              - - "2026-01-01T00:00:00Z"
                - "2026-01-01T23:59:59Z"
              - - "2026-01-02T00:00:00Z"
                - "2026-01-02T23:59:59Z"
  responses:
...
    ProvidersSearchResponse:
      description: Default Response
      content:
        application/json:
          schema:
            type: object
            properties:
              providers:
                description: List of providers matching the search criteria
                type: array
                items:
                  type: object
                  properties:
                    id:
                      description: The unique identifier for the provider
                      allOf:
                        - $ref: "#/components/schemas/Uuid"
                    first_name:
                      description: The provider's first name
                      type: string
                    last_name:
                      description: The provider's last name
                      type: string
                    profile_image_url:
                      description: URL to the provider's profile image
                      type: string
                    profile_bio:
                      description: A brief biography of the provider
                      type: string
                    genders:
                      description: The genders the provider identifies with
                      type: array
                      items:
                        type: string
                    races:
                      description: The races or ethnicities the provider identifies with
                      type: array
                      items:
                        type: string
                    slot_start_time:
                      description: The start time of the provider's earliest available slot.
                      type: string
                      format: date-time
                      pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
                  required:
                    - id
                    - first_name
                    - last_name
                  additionalProperties: false
            required:
              - providers
            additionalProperties: false
...
    ProvidersShowResponse:
      description: Default Response
      content:
        application/json:
          schema:
            type: object
            properties:
              id:
                description: The unique identifier for the provider
                allOf:
                  - $ref: "#/components/schemas/Uuid"
              first_name:
                description: The provider's first name
                type: string
              last_name:
                description: The provider's last name
                type: string
              profile_image_url:
                description: URL to the provider's profile image
                type: string
              profile_bio:
                description: A brief biography of the provider
                type: string
              profile_summary:
                description: Summary of the provider's profile
                type: string
              profile_approach:
                description: Introduction to potential patients, high level overview of approach
                  and background
                type: string
              profile_focus:
                description: Area of focus for provider. Helps align patient that best fits
                  provider
                type: string
              profile_journey:
                description: Provider's journey to become a therapist. Helps patients understand
                  motivations, signal of whether there's a fit in terms of
                  lived experience
                type: string
              profile_goals:
                description: Goals for patients, helps with patient alignment
                type: string
              profile_first_session:
                description: Expectations for the first session
                type: string
              role_string:
                description: String representation of the provider's role
                type: string
                enum:
                  - Therapist
                  - Prescriber
                  - Unknown
              insurances:
                description: List of insurances accepted by the provider
                type: array
                items:
                  type: string
              active_states:
                description: List of two-letter state codes for the states the provider is active in
                type: array
                items:
                  type: string
                  pattern: ^[A-Z]{2}$
              willing_to_see:
                description: Criteria of patients the provider is willing to see
                type: array
                items:
                  oneOf:
                    - const: adults
                      description: Adults (18+)
                    - const: children
                      description: Children (ages 5-7)
                    - const: minors
                      description: Minors (ages 13-17)
                    - const: preteen
                      description: Preteen (ages 8-12)
                    - const: individuals
                      description: Individual therapy
                    - const: couples
                      description: Couples therapy
                    - const: families
                      description: Family therapy
            required:
              - id
              - first_name
              - last_name
            additionalProperties: false
    ProvidersSlotsResponse:
      description: Default Response
      content:
        application/json:
          schema:
            type: object
            properties:
              slots:
                description: List of slots available for the provider
                type: array
                items:
                  type: object
                  properties:
                    provider_id:
                      description: The unique Rula-provided identifier for the provider
                      allOf:
                        - $ref: "#/components/schemas/Uuid"
                    duration_mins:
                      description: Duration of the slot in minutes
                      type: integer
                      exclusiveMinimum: 0
                      maximum: 9007199254740991
                      example: 45
                    location:
                      description: The location type of the slot
                      type: string
                      enum:
                        - in_person
                        - telemedicine
                    series_type:
                      description: Whether the slot is for a first appointment ("initial") or a subsequent session ("followup").
                      type: string
                      enum:
                        - initial
                        - followup
                    start_time_iso:
                      description: The ISO 8601 formatted start time of the slot
                      type: string
                      format: date-time
                      pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
                      example: 2024-01-15T10:00:00Z
                  required:
                    - provider_id
                    - duration_mins
                    - location
                    - series_type
                    - start_time_iso
                  additionalProperties: false
            required:
              - slots
            additionalProperties: false
  schemas:
...
    Uuid:
      type: string
      format: uuid
      pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
    RequestValidationError:
      type: object
      properties:
        message:
          type: string
          example: Invalid request
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                description: The field that failed validation
                example: querystring.uuid
              message:
                type: string
                description: Description of the validation failure
                example: must match pattern \"^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$\"
            required:
              - field
              - message
      required:
        - message
    InternalServerError:
      type: object
      properties:
        message:
          type: string
          example: Internal Server Error
        trace_id:
          type: string
          description: The trace ID for the error
      required:
        - message
...
  securitySchemes:
    auth0:    # <---- arbitrary name
      type: oauth2
      description: >-
        This API uses OAuth 2 with the client credentials flow.
        In addition to the endpoint-specific functional scope(s), every token must include
        the environment scope matching the target environment:
        `partner_api:production` for production, or `partner_api:staging` for staging.
      flows:
        clientCredentials:
          tokenUrl: https://login.rula.com/oauth/token
          refreshUrl: https://login.rula.com/oauth/token
          scopes:
            partner_api:insurances:read: Read access to insurance data
            partner_api:providers:read: Read access to provider data
            partner_api:patients:create: Create new patients
            partner_api:appointments:read: Read access to appointments data
            partner_api:appointments:create: Create new appointments
            partner_api:production: Access to the production environment
            partner_api:staging: Access to the staging environment